module-5-production-capstone/multi-agent-orchestrator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==1.12.0
httpx<0.28
python-dotenv==1.0.0
